+++ .github/workflows/ci.yml
@@ -61,7 +61,7 @@ jobs:
       matrix:
         python:
         - "3.8"  # oldest Python supported by PSF
-        - "3.9"  # newest Python that is stable
+        - "3.10"  # newest Python that is stable
         platform:
         - ubuntu-latest
         # - macos-latest
