### 2023-06-27 #5591 @phibeck
Hi @sgbaird, thanks for your submission. I'll be looking for reviewers next. A few things you can do to speed up the review process:

- it would be helpful if you could identify a few potential reviewers from this list https://reviewers.joss.theoj.org/reviewers (without tagging them with an @)
- there are some broken references in the acknowledgements. Please fix these and regenerate the pdf using `@editorialbot generate pdf`
- it looks like there are a couple of lists in your statement of need that aren't shown correctly, please check
- as a side note: your paper is on the longer side (we aim for <= 1000 words, but not strict), so you could think about moving some of the function documentation into the repository directly
