+++ reports/paper.bib
@@ -124,6 +124,23 @@
   file = {C\:\\Users\\sterg\\Zotero\\storage\\CLQCGW93\\Jain et al_2013_Commentary.pdf}
 }
 
+@article{ong_python_2013,
+  title = {Python {{Materials Genomics}} (Pymatgen): {{A}} Robust, Open-Source Python Library for Materials Analysis},
+  shorttitle = {Python {{Materials Genomics}} (Pymatgen)},
+  author = {Ong, Shyue Ping and Richards, William Davidson and Jain, Anubhav and Hautier, Geoffroy and Kocher, Michael and Cholia, Shreyas and Gunter, Dan and Chevrier, Vincent L. and Persson, Kristin A. and Ceder, Gerbrand},
+  year = {2013},
+  month = feb,
+  journal = {Computational Materials Science},
+  volume = {68},
+  pages = {314--319},
+  issn = {09270256},
+  doi = {10.1016/j.commatsci.2012.10.028},
+  urldate = {2022-06-09},
+  abstract = {We present the Python Materials Genomics (pymatgen) library, a robust, open-source Python library for materials analysis. A key enabler in high-throughput computational materials science efforts is a robust set of software tools to perform initial setup for the calculations (e.g., generation of structures and necessary input files) and post-calculation analysis to derive useful material properties from raw calculated data. The pymatgen library aims to meet these needs by (1) defining core Python objects for materials data representation, (2) providing a well-tested set of structure and thermodynamic analyses relevant to many applications, and (3) establishing an open platform for researchers to collaboratively develop sophisticated analyses of materials data obtained both from first principles calculations and experiments. The pymatgen library also provides convenient tools to obtain useful materials data via the Materials Project's REpresentational State Transfer (REST) Application Programming Interface (API). As an example, using pymatgen's interface to the Materials Project's RESTful API and phasediagram package, we demonstrate how the phase and electrochemical stability of a recently synthesized material, Li4SnS4, can be analyzed using a minimum of computing resources. We find that Li4SnS4 is a stable phase in the Li\textendash Sn\textendash S phase diagram (consistent with the fact that it can be synthesized), but the narrow range of lithium chemical potentials for which it is predicted to be stable would suggest that it is not intrinsically stable against typical electrodes used in lithium-ion batteries.},
+  langid = {english},
+  file = {C\:\\Users\\sterg\\Zotero\\storage\\GUGQ4GNV\\Ong et al. - 2013 - Python Materials Genomics (pymatgen) A robust, op.pdf}
+}
+
 @article{palizhati_agents_2022,
   title = {Agents for Sequential Learning Using Multiple-Fidelity Data},
   author = {Palizhati, Aini and Torrisi, Steven B. and Aykol, Muratahan and Suram, Santosh K. and Hummelsh{\o}j, Jens S. and Montoya, Joseph H.},
