### 2023-06-28 b2472f5
Update paper.md
https://github.com/openjournals/joss-reviews/issues/5591#issuecomment-1609650548
