+++ reports/paper.md
@@ -37,17 +37,17 @@ bibliography: paper.bib
 
 # Summary
 
-The progress of a machine learning field is both tracked and propelled through the development of robust benchmarks. While significant progress has been made to create standardized, easy-to-use benchmarks for molecular discovery (e.g., [@brown_guacamol_2019]), this remains a challenge for solid-state material discovery [@xie_crystal_2022; @zhao_physics_2022; @alverson_generative_2022]. To address this limitation, we propose `matbench_genmetrics`, an open-source Python library for benchmarking generative models for crystal structures. We incorporate benchmark datasets, splits, and metrics. The four evaluation metrics we use---validity, coverage, novelty, and uniqueness---are inspired by Guacamol [@brown_guacamol_2019] and Crystal Diffusion Variational AutoEncoder (CDVAE) [@xie_crystal_2022]. The evaluation metrics and benchmark classes are implemented in the `matbench-genmetrics.core` namespace package. The datasets and splitting are handled via the `matbench_genmetrics.mp_time_split` namespace package using Materials Project [@jain_commentary_2013] crystal structures and time-series cross-validation splits [@pedregosa_scikit-learn_2011] based on date first reported in literature, respectively. Future work may involve an automated leaderboard, an automated submission system, and easy-to-use examples for preparation and submission. We believe that `matbench-genmetrics` will provide the standardization and convenience required for rigorous benchmarking of crystal structure generative models. A visual overview of the `matbench_genmetrics` library is provided in \autoref{fig:summary}.
+The progress of a machine learning field is both tracked and propelled through the development of robust benchmarks. While significant progress has been made to create standardized, easy-to-use benchmarks for molecular discovery (e.g., [@brown_guacamol_2019]), this remains a challenge for solid-state material discovery [@xie_crystal_2022; @zhao_physics_2022; @alverson_generative_2022]. To address this limitation, we propose `matbench-genmetrics`, an open-source Python library for benchmarking generative models for crystal structures. We use four evaluation metrics inspired by Guacamol [@brown_guacamol_2019] and Crystal Diffusion Variational AutoEncoder (CDVAE) [@xie_crystal_2022]---validity, coverage, novelty, and uniqueness---to assess performance on Materials Project data splits using timeline-based cross-validation. We believe that `matbench-genmetrics` will provide the standardization and convenience required for rigorous benchmarking of crystal structure generative models. A visual overview of the `matbench-genmetrics` library is provided in \autoref{fig:summary}.
 
-![Summary visualization of `matbench_genmetrics` to evaluate crystal generative model performance using validity, coverage, novelty, and uniqueness metrics based on calendar-time splits of experimentally determined Materials Project database entries. Validity is the comparison of distribution characteristics (space group number) between the generated materials and the training and test sets. Coverage is the number of matches between the generated structures and a held-out test set. Novelty is a comparison between the generated and training structures. Finally, uniqueness is a measure of the number of repeats within the generated structures (i.e., comparing the set of generated structures to itself). For in-depth descriptions and equations for the four metrics described above, see https://matbench-genmetrics.readthedocs.io/en/latest/readme.html and https://github.com/sparks-baird/matbench-genmetrics/blob/main/reports/matbench-genmetrics.tex.\label{fig:summary}](figures/matbench-genmetrics.png)
[…]
