+++ reports/paper.md
@@ -45,7 +45,7 @@ The progress of a machine learning field is both tracked and propelled through t
 
 # Statement of need
 
-In materials informatics, benchmarks typically focus on property prediction using standard metrics like MAE and RMSE with random train-test splits [@dunn_benchmarking_2020]. However, crystal structure generative modeling lacks standardization compared to molecular generative modeling benchmarks. Molecular generative modeling benefits from widely adopted benchmark platforms like Guacamol [@brown_guacamol_2019] and Moses [@polykovskiy_molecular_2020], which offer easy installation, usage instructions, and leaderboards. In contrast, existing crystal structure generative modeling evaluations (e.g., CDVAE [@xie_crystal_2022], FTCP [@ren_invertible_2022], PGCGM [@zhao_physics_2022], CubicGAN [@zhao_high-throughput_2021], CrysTens [@alverson_generative_2022]) lack standardization, are challenging to install and apply to new models and datasets, and lack public leaderboards. While these evaluations are valuable for comparing models within their scope, a benchmarking platform is needed to promote standardization and enable robust comparisons.
+In materials informatics, benchmarks typically focus on property prediction using standard metrics like MAE and RMSE with random train-test splits [@dunn_benchmarking_2020]. However, crystal structure generative modeling lacks standardization compared to molecular generative modeling benchmarks. Molecular generative modeling efforts benefit from widely adopted benchmark platforms like Guacamol [@brown_guacamol_2019] and Moses [@polykovskiy_molecular_2020], which offer easy installation, usage instructions, and leaderboards. In contrast, existing crystal structure generative modeling evaluations (e.g., CDVAE [@xie_crystal_2022], FTCP [@ren_invertible_2022], PGCGM [@zhao_physics_2022], CubicGAN [@zhao_high-throughput_2021], CrysTens [@alverson_generative_2022]) lack standardization, are challenging to install and apply to new models and datasets, and lack public leaderboards. While these evaluations are valuable for comparing models within their scope, a benchmarking platform is needed to promote standardization and enable robust comparisons.
 
 In this work, we introduce `matbench-genmetrics`, a materials benchmarking platform for crystal structure generative models. We use concepts from molecular generative modeling benchmarking to create a set of evaluation metrics---validity, coverage, novelty, and uniqueness---which are broadly defined as follows:
 
