### 2023-06-28 32250c8
remove spurious tqdm import in readme
