### 2023-07-04 #5618 @editorialbot
[REVIEW]: matbench-genmetrics: A Python library for benchmarking crystal structure generative models using time-based splits of Materials Project structures
**Submitting author:** @sgbaird (<a href="http://orcid.org/0000-0002-4491-6876">Sterling Baird</a>)
**Repository:** https://github.com/sparks-baird/matbench-genmetrics
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.1
