### 2023-07-07 sparks-baird/matbench-genmetrics#80 @ml-evs
[JOSS review] Initial comments
Hi @sgbaird, just going to use this issue to track anything I spot on my first playaround with this:

- [ ] Incorrect link for the metrics description in the README
- [ ] Missing release v0.6.1 in GitHub (but present on PyPI)
- [ ] CHANGELOG.md has never been filled in (I guess it can just be removed)
- [ ] LICENSE.txt should probably use your full name rather than GitHub handle (though guess that's just preference)
- [ ] Unclear which dependencies are the "real" ones, `environment.yml` has much of it commented out (and has no pinned versions), `requirements.txt` is just a single line for a package not mentioned anywhere else, so I guess `setup.cfg` is the real one? I can't find any pinned dependencies anywhere which will hinder reproducibility going forward. This also seems to go against the Dependency Management section of the README
- [ ] The folders `configs`, `data`, `models`, `raw`, and `references` don't seem to contain anything useful
