### 2023-08-10 #5618 @phibeck
Thank you all for getting the review started! As you work through your checklists, please feel free to comment and ask questions in this thread. You are encouraged to create issues in the repository directly. When you do, please mention `openjournals/joss-reviews#5618` so that it creates a link in this thread and we can keep track of it.

Please let me know if you have any questions or if either of you requires some more time.
