### 2023-08-17 #5618 @mkhorton
Hi @phibeck, running through the checklist. Unfortunately I may have a conflict of interest: I have previously been on publications with the second author, Joseph Montoya. We worked together in the same research group until 2018 (i.e. outside the four year window in the COI policy), but [the most recent paper](https://pubs.acs.org/doi/abs/10.1021/acs.jcim.1c00340) we were on together actually only came out in 2021.
