### 2023-08-17 #5618 @danielskatz
If the work was done more then 4 years, but the paper appeared later, this is not a conflict for JOSS. The four years is about the collaborative relationship itself.
