### 2023-09-05 sparks-baird/matbench-genmetrics#81 @jamesrhester
Example in installation docs doesn't work
This is part of the JOSS review openjournals/joss-reviews#5618.

The example in the installation documentation doesn't work. After `pip install matbench-genmetrics` into a new Python 3.11 virtual environment, I have the following:

```
In [1]: from mp_time_split.utils.gen import DummyGenerator
Cell In[1], line 1
----> 1 from mp_time_split.utils.gen import DummyGenerator

ModuleNotFoundError: No module named 'mp_time_split'
```
which is presumably a missing module required for the example.

In any case, the next line also fails to work.

```
In [2]: from matbench_genmetrics.core import MPTSMetrics10, MPTSMetrics100, MPTSMetrics1000, MPTSMetrics10000
ImportError                               Traceback (most recent call last)
Cell In[2], line 1
----> 1 from matbench_genmetrics.core import MPTSMetrics10, MPTSMetrics100, MPTSMetrics1000, MPTSMetrics10000

ImportError: cannot import name 'MPTSMetrics10' from 'matbench_genmetrics.core' (/home/jrh/programs/JOSS/matbench-genmetrics/lib/python3.11/site-packages/matbench_genmetrics/core/__init__.py)
```
