### 2023-09-05 sparks-baird/matbench-genmetrics#81 @jamesrhester
Looks to be easily fixed by replacing the text with contents of `scripts/core/matbench_genmetrics_basic.py`
