### 2023-09-20 #5618 @ml-evs
Hi @phibeck, I have been continuing to list my comments at https://github.com/sparks-baird/matbench-genmetrics/issues/80. My main concern as raised by these comments is that while "complete" the library does not appear to have been "polished", there are lots of minor errors in the docs and defunct code paths/duplicated template code that seems to be leftover from project initialization (I can only find around 700 lines in total of active code in the main package). Coupled with the fact that there are no examples of "serious"/comparative benchmarks, i.e., benchmarking a real generative algorithm, I am struggling to extrapolate the package performance/functionality from the simple example provided. If my comments in https://github.com/sparks-baird/matbench-genmetrics/issues/80 could be addressed I would have an easier time recommending this for publication, but until then I find it hard to judge.
