### 2023-09-20 #5618 @ml-evs
Just to attempt to give a constructive suggestion, since actually performing generative ML is out of scope for JOSS, perhaps a nice self referential example would be to use the MP time split subpackage to "benchmark" our last n years of progress in materials discovery as a field, i.e., treat the slow march of time as a generative model. With a bit of fudging (depending on how the time split treats duplicates in the ICSD) I think this can test novelty, uniqueness and validity with more interpretable results than the given random model.
