+++ docs/metrics.md
@@ -1,6 +1,6 @@
 # Metrics
 
-The four metrics are validity, coverage, novelty, and uniqueness. We describe each in detail below.
+The four metrics used in `matbench-genmetrics` are validity, coverage, novelty, and uniqueness. We describe each in detail below.
 
 ## Validity
 
@@ -66,3 +66,7 @@ represent number of structures in the generated set, crystallographic distance a
 to `StructureMatcher` from `pymatgen.analysis.structure_matcher`, $i$-th
 structure of the generated set, $j$-th structure of the generated set, and a tolerance
 threshold, respectively.
+
+##
+
+While useful individually, these metrics can also be used as multi-criteria filtering. One reason this is important is because, standalone, the metrics can be "hacked" in some sense. For example, the novelty metric may be made perfect simply by generating a diverse set of nonsensical crystal structures. Likewise, the validity score may be bloated simply by passing in the list of training structures as the generated structures. To combat this, multiple criteria may be considered simultaneously: for example, requiring that a structure must be simultaneously novel, unique, and passing certain filtering criteria such as non-overlapping atoms, stoichiometry rules, or [checkCIF](https://checkcif.iucr.org/) criteria (DOI: [10.1107/S2056989019016244](https://dx.doi.org/10.1107/S2056989019016244)). Additional filters based on maching learning prediction models can be used for properties such as formation energy (i.e., must be negative), energy above hull, ICSD classification, and coordination number. Of particular interest is applying machine-learning based structural relaxation to the structures prior to filtering through universal interatomic potential models such as [M3GNet](https://doi.org/10.1038/s43588-022-00349-3).
