### 2023-12-08 9f149ed
File deletions, env, and dockerfile
https://github.com/sparks-baird/matbench-genmetrics/issues/80
