+++ environment.yml
@@ -7,54 +7,14 @@ channels:
 dependencies:
   - python>=3.6
   - pip
-  # BASICS
-  # - numpy
-  # - scipy
-  # - pandas
-  # - tqdm
-  # - click
-  # - scikit-learn
-  # - pymatgen
-  # - matminer
   - ipython
-  # VISUALIZATION
   - matplotlib
   - plotly
   - python-kaleido
-  # - ipympl  # interactive matplotlib plots
-  # - seaborn
-  # - altair
-  # - yellowbrick
-  # ML, STATS & DEEP LEARNING
-  # - statsmodels
-  # - scikit-learn
-  # - sktime
-  # - tslearn
-  # - xgboost
-  # - catboost
-  # - lightgbm
-  # - pytorch
-  # - fastai  # activate fastai channel above!
-  # - tensorflow
-  # - keras
-  # - spacy
-  # OTHER TOOLS
-  # - optuna
-  # - dask
-  # - snakeviz
   - pip:
      - -e .  # install git checkout of matbench-genmetrics in editable mode
-     #  - element-coder
-     # add here only pip-packages that are not available in conda/conda-forge! E.g.:
-     # - icecream
-     # - jax
-     # - numpyro
-     # - funsor
-     # - neptune-client
-     # - neptune-contrib
 
-  # DEVELOPMENT ONLY PACKAGES (could also be kept in a separate environment file)
-  - jupyterlab
+  # DEVELOPMENT ONLY PACKAGES
   - pytest
   - pytest-cov
   - tox
