### 2023-12-08 a5afdc2
remove tex files
These are represented in `docs/metrics.md` (https://matbench-genmetrics.readthedocs.io/en/latest/metrics.html)

`reports/equations/*.tex` were auto-generated from equations.nb, but then were manually copy-pasted into metrics.md.
