+++ reports/equations/validity.tex
@@ -1,4 +1,0 @@
-\begin{equation} \label{eq:validity}
-	1-\frac{w\left(\text{SG}_{\text{train}},\text{SG}_{\text{test}}\right)}{w\left(\text{SG}_{\text{train}},1\right)}
-\end{equation}
-where $\mathit{w}$, $\text{SG}_{\text{train}}$, and $\text{SG}_{\text{test}}$ represent Wasserstein distance, vector of space group numbers for the training data, and vector of space group numbers for the test data, respectively.
