+++ reports/matbench-genmetrics.tex
@@ -1,69 +1,0 @@
-\subsection{VALIDATION: Assessing Performance}
-\label{sec:validation}
-\textcolor{red}{****STERLING****}
-
-While significant progress has been made to create standardized, easy-to-use benchmarks for molecular discovery \cite{brownGuacaMolBenchmarkingModels2019}, this remains a challenge for solid-state materials \cite{spekCheckCIFValidationALERTS2020, xie_crystal_2022, zhao_physics_2022}. To address this limitation, we propose \texttt{matbench-genmetrics}, an open-source Python library for benchmarking generative models for crystal structures. We incorporate benchmark datasets, splits, and metrics inspired by Crystal Diffusion Variational AutoEncoder (CDVAE) \cite{xieCrystalDiffusionVariational2021}. We provide our own benchmarks using time-series style cross-validation splits from Materials Project via our \texttt{mp-time-split} package and we will also incorporate an automated leaderboard and submission system and provide an easy-to-use example for users to prepare and submit benchmarks for new models.
-
-Here we define the four metrics used in \texttt{matbench-genmetrics}: validity, coverage, novelty, and uniqueness (\cref{fig:matbench-genmetrics}).
-
-\begin{figure}
-	\centering
-	\includegraphics[width=0.48\textwidth]{sections/figs/metrics.png}
-	\caption{The four metrics of \texttt{matbench-genmetrics} for assessing performance of materials generative models are validity, coverage, novelty, and uniqueness. Validity is the comparison of distribution characteristics (space group number) between the generated materials and the training and test sets. Coverage is the number of matches between the generated structures and a held-out test set. Novelty is a comparison between the generated and training structures. Finally, uniqueness is a measure of the number of repeats within the generated structures (i.e., comparing the set of generated structures to itself).}
-	\label{fig:matbench-genmetrics}
-\end{figure}
-
-We define validity as one minus the Wasserstein distance between distribution of space group numbers for train and generated structures divided by the distance of the dummy case between train and the space group number 1:
-
-\begin{equation} \label{eq:validity}
-	1-\frac{w\left(\mathrm{SG}_{\mathrm{train}},\mathrm{SG}_{\mathrm{test}}\right)}{w\left(\mathrm{SG}_{\mathrm{train}},1\right)}
-\end{equation}
-where $w$, $\mathrm{SG}_{\mathrm{train}}$, and $\mathrm{SG}_{\mathrm{test}}$ represent Wasserstein distance, vector of space group numbers for the training data, and vector of space group numbers for the test data, respectively.
-
-Coverage (``predict the future'') is given by the match counts between the held-out test structures and the generated structures divided by the number of test structures:
-
-\begin{equation} \label{eq:coverage}
-	\frac{\sum _{i=1}^{n_{\text{test}}} \sum _{j=1}^{n_{\text{gen}}} \left(
-		\left\{
-		\begin{array}{cc}
-			1 & d\left(s_{\text{test},i},s_{\text{gen},j}\right)\leq \text{tol} \\
-			0 & d\left(s_{\text{test},i},s_{\text{gen},j}\right)>\text{tol} \\
-		\end{array}
-		\\
-		\right.
-		\right)}{n_{\text{test}}}
-\end{equation}
[…]
