### 2023-12-08 7ad1831
fix spurious link to .tex file
https://github.com/sparks-baird/matbench-genmetrics/issues/80
