### 2023-12-08 b04e9d3
fix metrics link in readme
https://github.com/sparks-baird/matbench-genmetrics/issues/80
