### 2023-12-08 f9e1e2a
massively simplify dockerfile
Install Python, run pip install command

https://github.com/sparks-baird/matbench-genmetrics/issues/80
