+++ tests/test_matbench_genmetrics.py
@@ -200,61 +200,3 @@ def test_cdvae_coverage():
 
 def test_cdvae_cov_struct_fingerprints():
     cdvae_cov_struct_fingerprints(dummy_structures, verbose=True)
-
-
-# %% Code Graveyard
-# def flatten_params(
-#     fixtures: List[Callable], expecteds: List[dict]
-# ) -> List[Tuple[Callable, str, npt.ArrayLike]]:
-#     """Match list of fixtures to list of dicts element-wise; create flat combinations.
-
-#     Based on `pytest nested function parametrization
-#     <https://stackoverflow.com/a/62033116/13697228>`_
-
-#     Parameters
-#     ----------
-#     fixtures : List[Callable]
-#         List of fixture functions.
-#     expecteds : List[dict]
-#         List of dictionaries mapping class variable names to expected values.
-
-#     Returns
-#     -------
-#         flat_fixtures : List[Callable]
-#             List of paired (repeated) fixture functions.
-#         flat_attributes : List[str]
-#             List of flattened class variable names.
-#         flat_values : List[Sequence]
-#             List of flattened expected values.
-
-#     Examples
-#     --------
-#     >>> @fixture
-#     >>> def dummy_gen_matcher():
-#     >>>    return GenMatcher(dummy_structures, dummy_structures)
-
-#     >>> @fixture
-#     >>> def dummy_gen_metrics():
-#     >>>    return GenMetrics(dummy_structures, dummy_structures)
-
-#     >>> fixtures = [dummy_gen_matcher, dummy_gen_metrics]
-#     >>> expecteds = [{"match_rate": [1.0]}, {"validity": [0.5], "novelty": [0.5], "uniqueness": [0.5]}] # noqa: E501
-#     >>> flatten_params(fixtures, expecteds)
-#     [(<function dummy_gen_matcher at 0x000001E1A89CED30>, 'match_rate', [1.0]),
-#     (<function dummy_gen_metrics at 0x000001E1A89CEF70>, 'validity', [0.5]),
-#     (<function dummy_gen_metrics at 0x000001E1A89CEF70>, 'novelty', [0.5]), (<function
-#     dummy_gen_metrics at 0x000001E1A89CEF70>, 'uniqueness', [0.5])]
-
-#     See also
-#     --------
-#     https://stackoverflow.com/a/42400786/13697228
-#     """
-#     combinations = zip(fixtures, expecteds)
-#     combinations: List[Tuple[Callable, str, npt.ArrayLike]] = []
-#     for fix, expected in zip(fixtures, expecteds):
-#         for attribute, value in expected.items():
-#             combinations.append((fix, attribute, value))
-#     return combinations
-
-
-# combinations = flatten_params(fixtures, expecteds)
