+++ tests/test_mp_time_split.py
@@ -76,14 +76,3 @@ if __name__ == "__main__":
     train_inputs, val_inputs, train_outputs, val_outputs = test_get_train_and_val_data()
     train_inputs, test_inputs, train_outputs, test_outputs = test_get_test_data()
     data
-
-# %% Code Graveyard
-# def test_data_snapshot():
-#     dummy_expt_df_check = load_dataframe_from_json(dummy_data_path)
-#     mpt = MPTimeSplit(num_sites=(1, 2), elements=["V"])
-#     dummy_expt_df = mpt.fetch_data()
-#     dummy_match = dummy_expt_df.compare(dummy_expt_df_check)
-#     if not dummy_match.empty:
-#         raise ValueError(
-#             f"dummy_expt_df and dummy_expt_df_check unmatched: {dummy_match}"
-#         )
