+++ setup.cfg
@@ -96,15 +96,6 @@ testing =
     mp-api
 
 [options.entry_points]
-# Add here console scripts like:
-# console_scripts =
-#     script_name = matbench_genmetrics.core.module:function
-# For example:
-# console_scripts =
-#     fibonacci = matbench_genmetrics.core.skeleton:run
-# And any other entry points, for example:
-# pyscaffold.cli =
-#     awesome = pyscaffoldext.awesome.extension:AwesomeExtension
 console_scripts =
     matbench-genmetrics = matbench_genmetrics.core.core:run
     matbench_genmetrics = matbench_genmetrics.core.core:run

[…]
