### 2023-12-08 3094805
add documentation
https://github.com/sparks-baird/matbench-genmetrics/issues/80
