+++ src/matbench_genmetrics/core/utils/match.py
@@ -9,6 +9,7 @@ from pymatgen.core.structure import Structure
 from scipy.spatial.distance import cdist, pdist, squareform
 from tqdm import tqdm
 from tqdm.notebook import tqdm as ipython_tqdm
+from typing import List
 
 warnings.filterwarnings(
     "ignore",
@@ -29,10 +30,22 @@ is_notebook = hasattr(__builtins__, "__IPYTHON__")
 
 
 def dummy_tqdm(x, **kwargs):  # noqa: E731
+    """
+    A dummy function that simply returns its input. Used as a placeholder for
+    the tqdm progress bar function when verbose mode is not enabled.
+    """
     return x
 
 
 def get_tqdm(verbose):
+    """
+    Returns the appropriate tqdm function based on the environment and verbosity. If verbose is False, returns a dummy function that does nothing.
+
+    Parameters
+    ----------
+    verbose : bool
+        If True, returns the tqdm function. If False, returns a dummy function.
+    """
     if verbose:
         return ipython_tqdm if is_notebook else tqdm
     else:
@@ -40,6 +53,24 @@ def get_tqdm(verbose):
 
 
 def structure_matcher(s1: Structure, s2: Structure):
+    """
+    Checks if two pymatgen Structure objects match according to pymatgen's
+    StructureMatcher criteria with the following settings:
+
+    StructureMatcher(stol=0.5, ltol=0.3, angle_tol=10.0)
+
+    Parameters
+    ----------
+    s1 : Structure
+        The first structure to compare.
+    s2 : Structure
+        The second structure to compare.
+
+    Returns
+    -------
+    bool
+        True if the structures match, False otherwise.
+    """
     return sm.fit(s1, s2)
 
 
@@ -47,12 +78,32 @@ pairwise_match_fn_dict = {"StructureMatcher": structure_matcher}
 
 
 def structure_pairwise_match_matrix(
-    test_structures,
-    gen_structures,
-    match_type="StructureMatcher",
-    verbose=False,
-    symmetric=False,
+    test_structures: List[Structure],
+    gen_structures: List[Structure],
+    match_type: str = "StructureMatcher",
+    verbose: bool = False,
+    symmetric: bool = False,
 ):
+    """
+    This function computes a pairwise match matrix between two lists of pymatgen
+    Structure objects using a specified match function. The match function is
+    determined by the `match_type` parameter.
+
+    Parameters
+    ----------
+    test_structures : List[Structure]
+        List of pymatgen Structure objects to be compared.
+    gen_structures : List[Structure]
+        List of pymatgen Structure objects to be compared.
+    match_type : str, optional
+        The type of match function to use, by default "StructureMatcher".
+    verbose : bool, optional
+        If True, the function will provide a running progress bar, by default
+        False.
+    symmetric : bool, optional
+        If True, the function will compute a symmetric match matrix else an
+        array in the style of cdist, by default False.
+    """
     # TODO: replace with group_structures to be faster
     pairwise_match_fn = pairwise_match_fn_dict[match_type]
     match_matrix = np.zeros((len(test_structures), len(gen_structures)))
@@ -75,11 +126,37 @@ bva = BVAnalyzer()
 
 
 def cdvae_cov_match_matrix(
-    test_fingerprints,
-    gen_fingerprints,
-    symmetric=False,
-    cutoff=10.0,
+    test_fingerprints: List[List[float]],
+    gen_fingerprints: List[List[float]],
+    symmetric: bool = False,
+    cutoff: float = 10.0,
 ):
+    """
+    This function computes a match matrix between two sets of fingerprints (test
+    and generated) based on a cutoff distance.
+
+    Parameters
+    ----------
+    test_fingerprints : List[List[float]]
+        List of test fingerprints to be compared. Each fingerprint is a list of
+        floats.
[…]
