### 2023-12-08 0118838
Update paper.md
https://github.com/openjournals/joss-reviews/issues/5618#issuecomment-1706026006
