+++ reports/paper.md
@@ -45,7 +45,9 @@ The progress of a machine learning field is both tracked and propelled through t
 
 # Statement of need
 
-In materials informatics, benchmarks typically focus on property prediction using standard metrics like MAE and RMSE with random train-test splits [@dunn_benchmarking_2020]. However, crystal structure generative modeling lacks standardization compared to molecular generative modeling benchmarks. Molecular generative modeling efforts benefit from widely adopted benchmark platforms like Guacamol [@brown_guacamol_2019] and Moses [@polykovskiy_molecular_2020], which offer easy installation, usage instructions, and leaderboards. In contrast, existing crystal structure generative modeling evaluations (e.g., benchmarks reported in CDVAE [@xie_crystal_2022], FTCP [@ren_invertible_2022], PGCGM [@zhao_physics_2022], CubicGAN [@zhao_high-throughput_2021], CrysTens [@alverson_generative_2022]) lack standardization, can be challenging to install and apply to new models and datasets, and lack public leaderboards. While these evaluations are valuable for comparing models within their scope, a benchmarking platform is needed to promote standardization and enable robust comparisons.
+In the field of materials informatics, where materials science intersects with machine learning, benchmarks play a crucial role in assessing model performance and enabling fair comparisons among various tools and models. Typically, these benchmarks focus on evaluating the accuracy of predictive models for materials properties, utilizing well-established metrics such as mean absolute error (MAE) and root-mean-square error (RMSE) to measure performance against actual measurements. A standard practice involves splitting the data into two parts, with one serving as training data for model development and the other as test data for assessing performance [@dunn_benchmarking_2020].
+
+However, benchmarking generative models, which aim to create entirely new data rather than focusing solely on predictive accuracy, presents unique challenges. While significant progress has been made in standardizing benchmarks for tasks like image generation and molecule synthesis, the field of crystal structure generative modeling lacks this level of standardization. Molecular generative modeling benefits from widely adopted benchmark platforms such as Guacamol [@brown_guacamol_2019] and Moses [@polykovskiy_molecular_2020], which offer easy installation, usage guidelines, and leaderboards for tracking progress. In contrast, existing evaluations in crystal structure generative modeling, as seen in CDVAE [@xie_crystal_2022], FTCP [@ren_invertible_2022], PGCGM [@zhao_physics_2022], CubicGAN [@zhao_high-throughput_2021], and CrysTens [@alverson_generative_2022], lack standardization, pose challenges in terms of installation and application to new models and datasets, and lack publicly accessible leaderboards. While these evaluations are valuable within their respective scopes, there is a clear need for a dedicated benchmarking platform to promote standardization and facilitate robust comparisons.
 
 In this work, we introduce `matbench-genmetrics`, a materials benchmarking platform for crystal structure generative models. We use concepts from molecular generative modeling benchmarking to create a set of evaluation metrics---validity, coverage, novelty, and uniqueness---which are broadly defined as follows:
 
@@ -77,6 +79,6 @@ We believe that the `matbench-genmetrics` ecosystem is a robust and easy-to-use 
 
 # Acknowledgements
 
-We acknowledge contributions from Kevin M. Jablonka ([\@kjappelbaum](https://github.com/kjappelbaum)), Matthew K. Horton ([\@mkhorton](https://github.com/mkhorton)), Kyle D. Miller ([\@kyledmiller](https://github.com/kyledmiller)), and Janosh Riebesell ([\@janosh](https://github.com/janosh)). S.G.B. and T.D.S. acknowledge support by the National Science Foundation, USA under Grant No. DMR-1651668.
[…]
