### 2023-12-08 94d00a8
ruamel<0.18
Noticed my scheduled tests stopped working, due to ruamel updating to 0.18 and deprecating safe_load.

matminer 0.9.0 uses safe_load, and a new release hasn't been added due to other errors: https://github.com/hackingmaterials/matminer/actions/runs/6823376825
