+++ setup.cfg
@@ -61,7 +61,7 @@ install_requires =
     matminer
     pybtex
     typing-extensions
-    ruamel<0.18
+    ruamel.yaml<0.18
     # smact
     # hydra-core
     # torch
