### 2023-12-08 26f54b7
don't match on energy_above_hull or formation_energy_per_atom
These values changed at some point between my snapshot and now
