### 2023-12-08 ec29a46
fix up error with input and output getting crossed
