### 2023-12-08 sparks-baird/matbench-genmetrics#80 @sgbaird
Addressed all points (can you verify the Dockerfile works now?) except for the "slow march of time" performance assessment which @hasan-sayeed will be taking point on.
