### 2023-12-08 #5618 @sgbaird
Hi @phibeck, sorry for such a delay.

@jamesrhester
> As a crystallographer but non ML specialist, I found the "Statement of Need" lacked context. Line 22 in the paper made no sense to me, which is not good for the first sentence. I would therefore like one or two further sentences added at the beginning explaining how ML uses benchmarks, e.g "in ML, the result of a prediction is evaluated using benchmarks, which are then used to adjust the ML weights. Typically, a crystal structure ML model has used benchmarks from...." (which might show I have no idea what I'm talking about). I think this will better help readers to quickly determine whether or not the paper is relevant to them.
>
> Once this is done I'm ready to sign off on my review.

I have updated this section in the text. Please let me know if this looks OK to you now!

In particular:

@ml-evs I have addressed all but one of the points in https://github.com/sparks-baird/matbench-genmetrics/issues/80. @hasan-sayeed is going to help me with testing out your "slow march of time" suggestion in https://github.com/openjournals/joss-reviews/issues/5618#issuecomment-1728516279.
