### 2023-12-11 #5618 @mkhorton
I'm happy to sign off on the text too, I'm just checking my remaining review items. This package is motivated by an important need in the community; I am not wholly satisfied that this solution is going to be sufficient as it stands today, but it's a good start and can be improved over time. The package also seems well thought through and structured (my compliments to the authors).

My apologies also for the delay with the review; I will be removing myself from the active reviewer list for a while since I don't want to normalize this sort of wait!
