### 2023-12-11 #5618 @mkhorton
To give additional context for my critical comment, it is not clear to me that (a) space group alone is sufficient to assess validity, or (b) what a good baseline distribution should be for space group. As far as I can tell, assessing the validity of a generated structure is an unsolved challenge and an active research need. Using the Materials Project as a baseline is a good start, but we might need to be more creative to find more comprehensive answers, and other people might reasonably prefer to use different baselines.

For this package to take off, I would also perhaps start actively soliciting suggestions for additional metrics to include. The desire to prevent metric hacking is also very important, and I hope to see more robust solutions here. The idea of including ML based pre-filters is useful, but also itself dependent on the quality of the ML filter: it could steer an otherwise good model awry, depending on implementation.

Some other points:

* Try the Sphinx `sphinx-prompt` plugin, to make your examples easier to copy-and-paste
* I am actually having trouble trying the example on your docs home page; can I verify it's correct and current? For example, I do not see `MPTSMetrics10` available to import from the `__init__.py`, I believe the example should look like:
```python
from tqdm import tqdm
from matbench_genmetrics.mp_time_split.utils.gen import DummyGenerator
from matbench_genmetrics.core.metrics import MPTSMetrics10, MPTSMetrics100, MPTSMetrics1000, MPTSMetrics10000
mptm = MPTSMetrics10(dummy=True)
for fold in mptm.folds:
    train_val_inputs = mptm.get_train_and_val_data(fold)
    dg = DummyGenerator()
    dg.fit(train_val_inputs)
    gen_structures = dg.gen(n=mptm.num_gen)
    mptm.evaluate_and_record(fold, gen_structures)
```
Unfortunately, `mptm.evaluate_and_record(fold, gen_structures)` seems to return `None` when I ran this, after printing information on `ElementProperty` and `SiteStatsFingerprint`.
* Likewise, I struggled to install via `pip` in a clean conda environment: I managed to install it in the end, but a pure pip install does not seem possible on the machine I tested it on (macOS 13.1, Python 3.9).
* Given that MP does sometimes deprecate older materials/structures over time, is there a plan to ensure that benchmarks generated with older versions will be compatible with newer versions (or at least some way to alert the user if not)?

If we can resolve the installation issue and double check and verify the code examples in the docs I'd be happy to sign off.
