### 2023-12-11 #5618 @jamesrhester
Regarding the example not working, I raised an issue a while back [here](https://github.com/sparks-baird/matbench-genmetrics/issues/81)
