### 2023-12-13 #5618 @ml-evs
I'm now happy with the state of the repository itself and the issues raised at https://github.com/sparks-baird/matbench-genmetrics/issues/80, with the one exception that I think the package still needs a good motivating example (as per my previous suggestion) -- I think this is already being worked on (and also ties in with what Matt (1) has said above). Another (non-blocking) suggestion (perhaps an easier one than using mp_time_split itself) would be to take the recently published [Gnome dataset](https://github.com/google-deepmind/materials_discovery) and compare that to MP via these metrics.

Pending this example case, the text is also good with me, modulo two minor comments

1. Figure 1 caption docs links are not being rendered as links
2. Since the initial submission, the [`matbench-discovery`](https://matbench-discovery.materialsproject.org/) package has also been released (with [preprint](https://arxiv.org/abs/2308.14920)). I think these two packages nicely complement each other, but I can see people getting confused (similar names, similar aims). Perhaps a sentence or two highlighting the differences would be helpful here?
