+++ README.md
@@ -48,7 +48,7 @@ change this to e.g. `python==3.8.*`. See [Advanced Installation](##Advanced-Inst
 > intended for fast prototyping and debugging, as it assumes only 10 generated structures.
 
 ```python
->>> from mp_time_split.utils.gen import DummyGenerator
+>>> from matbench_genmetrics.mp_time_split.utils.gen import DummyGenerator
 >>> from matbench_genmetrics.core.metrics import MPTSMetrics10, MPTSMetrics100, MPTSMetrics1000, MPTSMetrics10000
 >>> mptm = MPTSMetrics10(dummy=True)
 >>> for fold in mptm.folds:
