### 2024-01-26 #5618 @sgbaird
@phibeck sorry for the radio silence, @hasan-sayeed and I are actively working on the motivating example mentioned. Yes, that timeline is doable for us.
