### 2024-02-06 #5618 @hasan-sayeed
@ml-evs @sgbaird I've just added the motivating example mentioned above as a [notebook file](https://github.com/sparks-baird/matbench-genmetrics/blob/main/notebooks/core/2.0-matbench-genmetrics-materials_discovery_progress_benchmark.ipynb). This script treats materials discovered in the future as generated data and benchmarks the progress of materials discovery over time. It employs five-folds, each fold partitions the dataset into distinct training, validation, and generated sets using `mp_time_split` subpackage, simulating the incremental advancement of materials science. The result we got is as follows:

```
{0: {'coverage': 1.0,
     'novelty': 0.9444032616753151,
     'uniqueness': 0.9999354892341219,
     'validity': 0.8823073694046457},
 1: {'coverage': 1.0,
     'novelty': 0.9388435878428465,
     'uniqueness': 0.9998874668429091,
     'validity': 0.9657295385947532},
 2: {'coverage': 1.0,
     'novelty': 0.9267976278724981,
     'uniqueness': 0.9999187947547732,
     'validity': 0.9689630506461338},
 3: {'coverage': 1.0,
     'novelty': 0.9375463306152706,
     'uniqueness': 0.999984404745629,
     'validity': 0.942820170335924},
 4: {'coverage': 0.9996293550778355,
     'novelty': 0.9225352112676056,
     'uniqueness': 0.9999801452488405,
     'validity': 0.8635418248589397}}
```
