### 2024-02-07 #5618 @sgbaird
@hasan-sayeed you rock! Thank you. @ml-evs @mkhorton can you take a look and let me know if this looks ok to you in terms of a motivating example? If so, I'll move on and try to resolve all other outstanding issues. If not, let me know what else is needed.
