### 2024-03-19 #5618 @sgbaird
Hi all, I'm sorry for the trouble this has been. It wasn't my intention to draw this out so long. Between graduating, moving to a foreign country, a now three-month-old newborn, and taking a job where manuscripts and code contributions are not weighted directly as success metrics, it's been difficult to give the submission and those involved due justice in a reasonable timeframe. Things are starting to settle for me, but the timeline hasn't been fair to the reviewers or editors. Sorry 😞

I've done my best to address all points from the reviewers. Please have a look, and let me know if there are additional changes that need to be made or if I've missed something.

## @mkhorton

> Try the Sphinx sphinx-prompt plugin, to make your examples easier to copy-and-paste
Added
> I am actually having trouble trying the example on your docs home page; can I verify it's correct and current? For example, I do not see MPTSMetrics10 available to import from the __init__.py, I believe the example should look like:
Updated

> Unfortunately, `mptm.evaluate_and_record(fold, gen_structures)` seems to return `None` when I ran this, after printing information on `ElementProperty` and `SiteStatsFingerprint`.

It records it to an internal variable. If you think it would be better to also return it, I can update that. If so, would it be better to return the full dictionary or just the metrics associated with that fold? If it's fine as-is, I've updated the example with the `print` line directly after to make it clearer.

```python
>>> from matbench_genmetrics.mp_time_split.utils.gen import DummyGenerator
>>> from matbench_genmetrics.core.metrics import MPTSMetrics10, MPTSMetrics100, MPTSMetrics1000, MPTSMetrics10000
>>> mptm = MPTSMetrics10(dummy=True)
>>> for fold in mptm.folds:
>>>     train_val_inputs = mptm.get_train_and_val_data(fold)
>>>     dg = DummyGenerator()
>>>     dg.fit(train_val_inputs)
>>>     gen_structures = dg.gen(n=mptm.num_gen)
>>>     mptm.evaluate_and_record(fold, gen_structures)
>>> print(mptm.recorded_metrics)
```

> * Likewise, I struggled to install via `pip` in a clean conda environment: I managed to install it in the end, but a pure pip install does not seem possible on the machine I tested it on (macOS 13.1, Python 3.9).

I think this should be resolved now. I don't own a Mac, but it seems OK on Windows and Linux. One of the difficult packages in terms of dependencies was `pyxtal`, which I was only using lightly for testing, so I replaced it with some custom code.

> Given that MP does sometimes deprecate older materials/structures over time, is there a plan to ensure that benchmarks generated with older versions will be compatible with newer versions (or at least some way to alert the user if not)?

I've given this some thought, but I don't have a good solution in mind. I think alerting the user to any potential differences makes sense. The datasets are stored on FigShare, so that helps with reproducibility. Since the benchmarks may take a long time to

## @ml-evs

> Figure 1 caption docs links are not being rendered as links

Updated

> 2\. Since the initial submission, the [`matbench-discovery`](https://matbench-discovery.materialsproject.org/) package has also been released (with [preprint](https://arxiv.org/abs/2308.14920)). I think these two packages nicely complement each other, but I can see people getting confused (similar names, similar aims). Perhaps a sentence or two highlighting the differences would be helpful here?

Added to the README and now cited in manuscript

> The example looks good but needs more explanation; for example, am I meant to be mentally plotting how novelty changes across the folds pasted above? I also couldn't see it mentioned in the docs or website when I last looked (and there have been no further commits).

Added explanations and a figure. Lmk if OK or needs changes.

## @jamesrhester

> Regarding the example not working, I raised an issue a while back [here](https://github.com/sparks-baird/matbench-genmetrics/issues/81)

This should be resolved now.

##

Thank you to everyone for your patience and feedback.
