### 2024-03-20 #5618 @ml-evs
Thanks for the updates @sgbaird and for the extra legwork. I'm now happy to recommend this for acceptance and have completed my checklist above. The package is well-developed, well-documented, and well-situated in the context of the field. The MP time split example results are interesting and I would like to give them more thought, but I don't think that should hold up this paper if everyone else is happy.
