### 2024-04-02 #5618 @mkhorton
I'm also happy that my questions have been responded to, thank you @sgbaird and @phibeck.

> If it's fine as-is, I've updated the example with the print line directly after to make it clearer.

This is sufficient, it was just unclear from the example.

> The datasets are stored on FigShare, so that helps with reproducibility.

This should be sufficient too. I think this is a larger community problem, so I don't want to hold up this specific review due to this issue.

>  Between graduating, moving to a foreign country, a now three-month-old newborn, and taking a job where manuscripts and code contributions are not weighted directly as success metrics, it's been difficult to give the submission and those involved due justice in a reasonable timeframe. Things are starting to settle for me, but the timeline hasn't been fair to the reviewers or editors.

No need to apologise @sgbaird, at least not to myself as reviewer, we're all trying our best! It's been difficult for me to find time to review too. Congratulations on the newborn :)

I will need to do a final re-try of the installation before I can sign off the last remaining items.
