### 2024-05-01 #5618 @mkhorton
I do want to echo my previous comment however: this package is a great start and has a robust statement of need, however I do think there need to be better metrics included, especially for validity, but also questions of novelty can be subtle too (for example, the case of ordered approximations). I have no doubt the authors will continue to improve the package over time and support its publication, but I hope additional metrics might be considered -- this might even be a good area for collaboration within the broader community.
