### 2024-05-01 #5618 @sgbaird
> Thank you @mkhorton for finishing your review and for your evaluation and recommendation for future development. Thank you, @ml-evs, @mkhorton, and @jamesrhester for your thorough reviews and for your patience!
>
> @sgbaird the reviewers have recommended the submission for publication. There are a few more steps before we finalize the publication. At this point could you please:

> * [x]  Make a tagged release of your software, and list the version tag of the archived version here.

I released [v0.6.5](https://github.com/sparks-baird/matbench-genmetrics/tree/v0.6.5) a bit ago, and there are no further commits so far.

> * [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

> * [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

![image](https://github.com/openjournals/joss-reviews/assets/45469701/18a6f2e0-0b53-44d3-925d-4c7460ce00a1)

![image](https://github.com/openjournals/joss-reviews/assets/45469701/2688b97a-51a4-45db-943a-afd72101f855)

> * [x]  Please list the DOI of the archived version here.

[10.5281/zenodo.10840604](https://doi.org/10.5281/zenodo.10840604)

Let me know if you need anything else from my end!
