### 2024-05-09 #5618 @sgbaird
> @sgbaird thanks! Here are a few more comments/suggestions for the manuscript. Please have a look when you have a moment.
>
> * line 12: I would omit the square brackets since the citation seems part of the sentence

Done!

> * line 27: the acronyms aren't used in the remainder of the manuscript, so it doesn't seem necessary to introduce them here

Agreed, removed

> * line 57: perhaps you could introduce a linebreak in front of `matbench_genmetrics.core` since it exceeds the linewidth (not sure why this isn't done automatically..)

> * line 165ff: it seems this reference has been published in the meantime, please update the reference accordingly ([nature.com/articles/s41524-023-00987-9](https://www.nature.com/articles/s41524-023-00987-9))

Updated!

> Since I cannot find a record of the last two co-authors' contribution in your repository, could you please state their contributions here for the record of review? You can check out the guidance for authorship here: [joss.readthedocs.io/en/latest/submitting.html#authorship](https://joss.readthedocs.io/en/latest/submitting.html#authorship) Thanks!

@JosephMontoya-TRI supplied code and an implementation related to the `mp-time-split` package.
@sp8rks participated in the ideation and development / funding, vision, etc.
