### 2024-05-25 #5618 @Kevin-Mattheus-Moerman
@sgbaird as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, I only have the below points that require your attention:

- [ ] Check spelling for `practioners` this should perhaps be `practitioners`
- [ ] On the use of ChatGPT can you please clarify how it was used here in more detail?
