### 2024-05-25 #5618 @sgbaird
> @sgbaird as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, I only have the below points that require your attention:
>
> * [ ]  Check spelling for `practioners` this should perhaps be `practitioners`
> * [ ]  In your affiliations, please spell out `USA` as `United States of America`.
> * [ ]  In the 3rd affiliation, please add the country (and you may remove the zip code if you like, this is not needed).
> * [ ]  For the reference "JARVIS-Leaderboard: A Large Scale Benchmark of Materials Design Methods", you cite an Arxiv link, if you think it is appropriate, you could instead refer to the version which, if I'm not mistaken, now appears published here: [doi.org/10.1038/s41524-024-01259-w](https://doi.org/10.1038/s41524-024-01259-w)
> * [ ]  On the use of ChatGPT can you please clarify how it was used here in more detail?

Hi, I think all of these are addressed now. Can you take a look?
