### 2024-05-27 #5618 @sgbaird
> @sgbaird congratulations on this JOSS publication!
>
> Thanks for editing @phibeck !
>
> And a special thank you to the reviewers: @ml-evs, @mkhorton, @jamesrhester !!!!

Amazing! Thank you all so much for the support and patience 🤗 I really appreciate your efforts!
